Implement the RollCall api
Summary:
This allows an application to listen to all subscriptions and
unsubscription requests of the system.

The RocketEngine aggregator will be part of a separate patch.

Test Plan: unit test attached.

Reviewers: stupaq-fb, pja

Reviewed By: pja

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D35097

+++ src/messages/event_loop.cc
@@ -664,7 +664,7 @@ EventLoop::Run() {
     memset(&sin, 0, sizeof(sin));
     sin.sin6_family = AF_INET6;
     sin.sin6_addr = in6addr_any;
-    sin.sin6_port = htons(port_number_);
+    sin.sin6_port = htons(static_cast<uint16_t>(port_number_));
 
     // Create libevent connection listener.
     listener_ = evconnlistener_new_bind(
